+++ yunjian-20211217/alibabacloud_yunjian20211217/client.py
@@ -47,6 +47,12 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.CreateDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.CreateDemandPlanResponse:
+        """
+        @param request: CreateDemandPlanRequest
+        @param headers: CreateDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.user_id):
@@ -62,6 +68,8 @@ class Client(OpenApiClient):
             body['period'] = request.period
         if not UtilClient.is_unset(request.source):
             body['source'] = request.source
+        if not UtilClient.is_unset(request.target_cid):
+            body['targetCid'] = request.target_cid
         if not UtilClient.is_unset(request.type):
             body['type'] = request.type
         real_headers = {}
@@ -96,6 +104,12 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.CreateDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.CreateDemandPlanResponse:
+        """
+        @param request: CreateDemandPlanRequest
+        @param headers: CreateDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.user_id):
@@ -111,6 +125,8 @@ class Client(OpenApiClient):
             body['period'] = request.period
         if not UtilClient.is_unset(request.source):
             body['source'] = request.source
+        if not UtilClient.is_unset(request.target_cid):
+            body['targetCid'] = request.target_cid
         if not UtilClient.is_unset(request.type):
             body['type'] = request.type
         real_headers = {}
@@ -143,6 +159,10 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.CreateDemandPlanRequest,
     ) -> yunjian_20211217_models.CreateDemandPlanResponse:
+        """
+        @param request: CreateDemandPlanRequest
+        @return: CreateDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.CreateDemandPlanHeaders()
         return self.create_demand_plan_with_options(request, headers, runtime)
@@ -151,6 +171,10 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.CreateDemandPlanRequest,
     ) -> yunjian_20211217_models.CreateDemandPlanResponse:
+        """
+        @param request: CreateDemandPlanRequest
+        @return: CreateDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.CreateDemandPlanHeaders()
         return await self.create_demand_plan_with_options_async(request, headers, runtime)
@@ -161,6 +185,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.CreateDemandPlanV2Headers,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.CreateDemandPlanV2Response:
+        """
+        @summary 创建plan2.0版本
+        
+        @param request: CreateDemandPlanV2Request
+        @param headers: CreateDemandPlanV2Headers
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateDemandPlanV2Response
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.user_id):
@@ -174,6 +206,8 @@ class Client(OpenApiClient):
             body['name'] = request.name
         if not UtilClient.is_unset(request.product_type):
             body['productType'] = request.product_type
+        if not UtilClient.is_unset(request.target_cid):
+            body['targetCid'] = request.target_cid
         if not UtilClient.is_unset(request.type):
             body['type'] = request.type
         real_headers = {}
@@ -208,6 +242,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.CreateDemandPlanV2Headers,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.CreateDemandPlanV2Response:
+        """
+        @summary 创建plan2.0版本
+        
+        @param request: CreateDemandPlanV2Request
+        @param headers: CreateDemandPlanV2Headers
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: CreateDemandPlanV2Response
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.user_id):
@@ -221,6 +263,8 @@ class Client(OpenApiClient):
             body['name'] = request.name
         if not UtilClient.is_unset(request.product_type):
             body['productType'] = request.product_type
+        if not UtilClient.is_unset(request.target_cid):
+            body['targetCid'] = request.target_cid
         if not UtilClient.is_unset(request.type):
             body['type'] = request.type
         real_headers = {}
@@ -253,6 +297,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.CreateDemandPlanV2Request,
     ) -> yunjian_20211217_models.CreateDemandPlanV2Response:
+        """
+        @summary 创建plan2.0版本
+        
+        @param request: CreateDemandPlanV2Request
+        @return: CreateDemandPlanV2Response
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.CreateDemandPlanV2Headers()
         return self.create_demand_plan_v2with_options(request, headers, runtime)
@@ -261,6 +311,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.CreateDemandPlanV2Request,
     ) -> yunjian_20211217_models.CreateDemandPlanV2Response:
+        """
+        @summary 创建plan2.0版本
+        
+        @param request: CreateDemandPlanV2Request
+        @return: CreateDemandPlanV2Response
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.CreateDemandPlanV2Headers()
         return await self.create_demand_plan_v2with_options_async(request, headers, runtime)
@@ -271,6 +327,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeleteUrgentDemandItemHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeleteUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单ite 删除
+        
+        @param request: DeleteUrgentDemandItemRequest
+        @param headers: DeleteUrgentDemandItemHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeleteUrgentDemandItemResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.id):
@@ -308,6 +372,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeleteUrgentDemandItemHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeleteUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单ite 删除
+        
+        @param request: DeleteUrgentDemandItemRequest
+        @param headers: DeleteUrgentDemandItemHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeleteUrgentDemandItemResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.id):
@@ -343,6 +415,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeleteUrgentDemandItemRequest,
     ) -> yunjian_20211217_models.DeleteUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单ite 删除
+        
+        @param request: DeleteUrgentDemandItemRequest
+        @return: DeleteUrgentDemandItemResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeleteUrgentDemandItemHeaders()
         return self.delete_urgent_demand_item_with_options(request, headers, runtime)
@@ -351,6 +429,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeleteUrgentDemandItemRequest,
     ) -> yunjian_20211217_models.DeleteUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单ite 删除
+        
+        @param request: DeleteUrgentDemandItemRequest
+        @return: DeleteUrgentDemandItemResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeleteUrgentDemandItemHeaders()
         return await self.delete_urgent_demand_item_with_options_async(request, headers, runtime)
@@ -361,6 +445,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeleteUrgentDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeleteUrgentDemandPlanResponse:
+        """
+        @summary 紧急需求单plan删除
+        
+        @param request: DeleteUrgentDemandPlanRequest
+        @param headers: DeleteUrgentDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeleteUrgentDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.id):
@@ -398,6 +490,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeleteUrgentDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeleteUrgentDemandPlanResponse:
+        """
+        @summary 紧急需求单plan删除
+        
+        @param request: DeleteUrgentDemandPlanRequest
+        @param headers: DeleteUrgentDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeleteUrgentDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.id):
@@ -433,6 +533,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeleteUrgentDemandPlanRequest,
     ) -> yunjian_20211217_models.DeleteUrgentDemandPlanResponse:
+        """
+        @summary 紧急需求单plan删除
+        
+        @param request: DeleteUrgentDemandPlanRequest
+        @return: DeleteUrgentDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeleteUrgentDemandPlanHeaders()
         return self.delete_urgent_demand_plan_with_options(request, headers, runtime)
@@ -441,6 +547,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeleteUrgentDemandPlanRequest,
     ) -> yunjian_20211217_models.DeleteUrgentDemandPlanResponse:
+        """
+        @summary 紧急需求单plan删除
+        
+        @param request: DeleteUrgentDemandPlanRequest
+        @return: DeleteUrgentDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeleteUrgentDemandPlanHeaders()
         return await self.delete_urgent_demand_plan_with_options_async(request, headers, runtime)
@@ -451,6 +563,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeliveryItemDetailSynHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeliveryItemDetailSynResponse:
+        """
+        @summary 交付信息同步
+        
+        @param request: DeliveryItemDetailSynRequest
+        @param headers: DeliveryItemDetailSynHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeliveryItemDetailSynResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.channel):
@@ -492,6 +612,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.DeliveryItemDetailSynHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.DeliveryItemDetailSynResponse:
+        """
+        @summary 交付信息同步
+        
+        @param request: DeliveryItemDetailSynRequest
+        @param headers: DeliveryItemDetailSynHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: DeliveryItemDetailSynResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.channel):
@@ -531,6 +659,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeliveryItemDetailSynRequest,
     ) -> yunjian_20211217_models.DeliveryItemDetailSynResponse:
+        """
+        @summary 交付信息同步
+        
+        @param request: DeliveryItemDetailSynRequest
+        @return: DeliveryItemDetailSynResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeliveryItemDetailSynHeaders()
         return self.delivery_item_detail_syn_with_options(request, headers, runtime)
@@ -539,6 +673,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.DeliveryItemDetailSynRequest,
     ) -> yunjian_20211217_models.DeliveryItemDetailSynResponse:
+        """
+        @summary 交付信息同步
+        
+        @param request: DeliveryItemDetailSynRequest
+        @return: DeliveryItemDetailSynResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.DeliveryItemDetailSynHeaders()
         return await self.delivery_item_detail_syn_with_options_async(request, headers, runtime)
@@ -549,6 +689,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandItemListHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandItemListResponse:
+        """
+        @summary 查询报备单中报备项列表
+        
+        @param request: GetUrgentDemandItemListRequest
+        @param headers: GetUrgentDemandItemListHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandItemListResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.commodity_code):
@@ -596,6 +744,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandItemListHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandItemListResponse:
+        """
+        @summary 查询报备单中报备项列表
+        
+        @param request: GetUrgentDemandItemListRequest
+        @param headers: GetUrgentDemandItemListHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandItemListResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.commodity_code):
@@ -641,6 +797,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandItemListRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandItemListResponse:
+        """
+        @summary 查询报备单中报备项列表
+        
+        @param request: GetUrgentDemandItemListRequest
+        @return: GetUrgentDemandItemListResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandItemListHeaders()
         return self.get_urgent_demand_item_list_with_options(request, headers, runtime)
@@ -649,6 +811,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandItemListRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandItemListResponse:
+        """
+        @summary 查询报备单中报备项列表
+        
+        @param request: GetUrgentDemandItemListRequest
+        @return: GetUrgentDemandItemListResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandItemListHeaders()
         return await self.get_urgent_demand_item_list_with_options_async(request, headers, runtime)
@@ -659,6 +827,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandPlanDetailHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanDetailResponse:
+        """
+        @summary getUrgentDemandPlanDetail
+        
+        @param request: GetUrgentDemandPlanDetailRequest
+        @param headers: GetUrgentDemandPlanDetailHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandPlanDetailResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.plan_id):
@@ -694,6 +870,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandPlanDetailHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanDetailResponse:
+        """
+        @summary getUrgentDemandPlanDetail
+        
+        @param request: GetUrgentDemandPlanDetailRequest
+        @param headers: GetUrgentDemandPlanDetailHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandPlanDetailResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.plan_id):
@@ -727,6 +911,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandPlanDetailRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanDetailResponse:
+        """
+        @summary getUrgentDemandPlanDetail
+        
+        @param request: GetUrgentDemandPlanDetailRequest
+        @return: GetUrgentDemandPlanDetailResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandPlanDetailHeaders()
         return self.get_urgent_demand_plan_detail_with_options(request, headers, runtime)
@@ -735,6 +925,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandPlanDetailRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanDetailResponse:
+        """
+        @summary getUrgentDemandPlanDetail
+        
+        @param request: GetUrgentDemandPlanDetailRequest
+        @return: GetUrgentDemandPlanDetailResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandPlanDetailHeaders()
         return await self.get_urgent_demand_plan_detail_with_options_async(request, headers, runtime)
@@ -745,6 +941,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandPlanListHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanListResponse:
+        """
+        @summary 查询报备单列表
+        
+        @param request: GetUrgentDemandPlanListRequest
+        @param headers: GetUrgentDemandPlanListHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandPlanListResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.current):
@@ -788,6 +992,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.GetUrgentDemandPlanListHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanListResponse:
+        """
+        @summary 查询报备单列表
+        
+        @param request: GetUrgentDemandPlanListRequest
+        @param headers: GetUrgentDemandPlanListHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetUrgentDemandPlanListResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.current):
@@ -829,6 +1041,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandPlanListRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanListResponse:
+        """
+        @summary 查询报备单列表
+        
+        @param request: GetUrgentDemandPlanListRequest
+        @return: GetUrgentDemandPlanListResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandPlanListHeaders()
         return self.get_urgent_demand_plan_list_with_options(request, headers, runtime)
@@ -837,6 +1055,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.GetUrgentDemandPlanListRequest,
     ) -> yunjian_20211217_models.GetUrgentDemandPlanListResponse:
+        """
+        @summary 查询报备单列表
+        
+        @param request: GetUrgentDemandPlanListRequest
+        @return: GetUrgentDemandPlanListResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.GetUrgentDemandPlanListHeaders()
         return await self.get_urgent_demand_plan_list_with_options_async(request, headers, runtime)
@@ -847,6 +1071,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.PushResourcePlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.PushResourcePlanResponse:
+        """
+        @summary ecs资源方案
+        
+        @param request: PushResourcePlanRequest
+        @param headers: PushResourcePlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: PushResourcePlanResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.buffer_cnt):
@@ -894,6 +1126,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.PushResourcePlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.PushResourcePlanResponse:
+        """
+        @summary ecs资源方案
+        
+        @param request: PushResourcePlanRequest
+        @param headers: PushResourcePlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: PushResourcePlanResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.buffer_cnt):
@@ -939,6 +1179,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.PushResourcePlanRequest,
     ) -> yunjian_20211217_models.PushResourcePlanResponse:
+        """
+        @summary ecs资源方案
+        
+        @param request: PushResourcePlanRequest
+        @return: PushResourcePlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.PushResourcePlanHeaders()
         return self.push_resource_plan_with_options(request, headers, runtime)
@@ -947,6 +1193,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.PushResourcePlanRequest,
     ) -> yunjian_20211217_models.PushResourcePlanResponse:
+        """
+        @summary ecs资源方案
+        
+        @param request: PushResourcePlanRequest
+        @return: PushResourcePlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.PushResourcePlanHeaders()
         return await self.push_resource_plan_with_options_async(request, headers, runtime)
@@ -957,6 +1209,14 @@ class Client(OpenApiClient):
         headers: Dict[str, str],
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.QueryDeliveredSupplyItemsResponse:
+        """
+        @summary 查询accountId下所有存在交付状态（包括部分）的报备数据, 以及开通数据情况
+        
+        @param request: QueryDeliveredSupplyItemsRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: QueryDeliveredSupplyItemsResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.account_id):
@@ -989,6 +1249,14 @@ class Client(OpenApiClient):
         headers: Dict[str, str],
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.QueryDeliveredSupplyItemsResponse:
+        """
+        @summary 查询accountId下所有存在交付状态（包括部分）的报备数据, 以及开通数据情况
+        
+        @param request: QueryDeliveredSupplyItemsRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: QueryDeliveredSupplyItemsResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.account_id):
@@ -1019,6 +1287,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.QueryDeliveredSupplyItemsRequest,
     ) -> yunjian_20211217_models.QueryDeliveredSupplyItemsResponse:
+        """
+        @summary 查询accountId下所有存在交付状态（包括部分）的报备数据, 以及开通数据情况
+        
+        @param request: QueryDeliveredSupplyItemsRequest
+        @return: QueryDeliveredSupplyItemsResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = {}
         return self.query_delivered_supply_items_with_options(request, headers, runtime)
@@ -1027,6 +1301,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.QueryDeliveredSupplyItemsRequest,
     ) -> yunjian_20211217_models.QueryDeliveredSupplyItemsResponse:
+        """
+        @summary 查询accountId下所有存在交付状态（包括部分）的报备数据, 以及开通数据情况
+        
+        @param request: QueryDeliveredSupplyItemsRequest
+        @return: QueryDeliveredSupplyItemsResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = {}
         return await self.query_delivered_supply_items_with_options_async(request, headers, runtime)
@@ -1037,6 +1317,14 @@ class Client(OpenApiClient):
         headers: Dict[str, str],
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.QueryPeriodBudgetBillResponse:
+        """
+        @summary 查询账单预算数据
+        
+        @param request: QueryPeriodBudgetBillRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: QueryPeriodBudgetBillResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.object_ids):
@@ -1071,6 +1359,14 @@ class Client(OpenApiClient):
         headers: Dict[str, str],
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.QueryPeriodBudgetBillResponse:
+        """
+        @summary 查询账单预算数据
+        
+        @param request: QueryPeriodBudgetBillRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: QueryPeriodBudgetBillResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.object_ids):
@@ -1103,6 +1399,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.QueryPeriodBudgetBillRequest,
     ) -> yunjian_20211217_models.QueryPeriodBudgetBillResponse:
+        """
+        @summary 查询账单预算数据
+        
+        @param request: QueryPeriodBudgetBillRequest
+        @return: QueryPeriodBudgetBillResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = {}
         return self.query_period_budget_bill_with_options(request, headers, runtime)
@@ -1111,6 +1413,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.QueryPeriodBudgetBillRequest,
     ) -> yunjian_20211217_models.QueryPeriodBudgetBillResponse:
+        """
+        @summary 查询账单预算数据
+        
+        @param request: QueryPeriodBudgetBillRequest
+        @return: QueryPeriodBudgetBillResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = {}
         return await self.query_period_budget_bill_with_options_async(request, headers, runtime)
@@ -1121,6 +1429,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.SaveUrgentDemandItemHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.SaveUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单item新增
+        
+        @param request: SaveUrgentDemandItemRequest
+        @param headers: SaveUrgentDemandItemHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: SaveUrgentDemandItemResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.account_id):
@@ -1180,6 +1496,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.SaveUrgentDemandItemHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.SaveUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单item新增
+        
+        @param request: SaveUrgentDemandItemRequest
+        @param headers: SaveUrgentDemandItemHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: SaveUrgentDemandItemResponse
+        """
         UtilClient.validate_model(request)
         body = {}
         if not UtilClient.is_unset(request.account_id):
@@ -1237,6 +1561,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.SaveUrgentDemandItemRequest,
     ) -> yunjian_20211217_models.SaveUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单item新增
+        
+        @param request: SaveUrgentDemandItemRequest
+        @return: SaveUrgentDemandItemResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.SaveUrgentDemandItemHeaders()
         return self.save_urgent_demand_item_with_options(request, headers, runtime)
@@ -1245,6 +1575,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.SaveUrgentDemandItemRequest,
     ) -> yunjian_20211217_models.SaveUrgentDemandItemResponse:
+        """
+        @summary 紧急需求单item新增
+        
+        @param request: SaveUrgentDemandItemRequest
+        @return: SaveUrgentDemandItemResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.SaveUrgentDemandItemHeaders()
         return await self.save_urgent_demand_item_with_options_async(request, headers, runtime)
@@ -1255,6 +1591,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.SubmitUrgentDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.SubmitUrgentDemandPlanResponse:
+        """
+        @summary submitUrgentDemandPlan
+        
+        @param request: SubmitUrgentDemandPlanRequest
+        @param headers: SubmitUrgentDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: SubmitUrgentDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.plan_id):
@@ -1292,6 +1636,14 @@ class Client(OpenApiClient):
         headers: yunjian_20211217_models.SubmitUrgentDemandPlanHeaders,
         runtime: util_models.RuntimeOptions,
     ) -> yunjian_20211217_models.SubmitUrgentDemandPlanResponse:
+        """
+        @summary submitUrgentDemandPlan
+        
+        @param request: SubmitUrgentDemandPlanRequest
+        @param headers: SubmitUrgentDemandPlanHeaders
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: SubmitUrgentDemandPlanResponse
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.plan_id):
@@ -1327,6 +1679,12 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.SubmitUrgentDemandPlanRequest,
     ) -> yunjian_20211217_models.SubmitUrgentDemandPlanResponse:
+        """
+        @summary submitUrgentDemandPlan
+        
+        @param request: SubmitUrgentDemandPlanRequest
+        @return: SubmitUrgentDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.SubmitUrgentDemandPlanHeaders()
         return self.submit_urgent_demand_plan_with_options(request, headers, runtime)
@@ -1335,6 +1693,124 @@ class Client(OpenApiClient):
         self,
         request: yunjian_20211217_models.SubmitUrgentDemandPlanRequest,
     ) -> yunjian_20211217_models.SubmitUrgentDemandPlanResponse:
+        """
+        @summary submitUrgentDemandPlan
+        
+        @param request: SubmitUrgentDemandPlanRequest
+        @return: SubmitUrgentDemandPlanResponse
+        """
         runtime = util_models.RuntimeOptions()
         headers = yunjian_20211217_models.SubmitUrgentDemandPlanHeaders()
         return await self.submit_urgent_demand_plan_with_options_async(request, headers, runtime)
+
+    def accept_fulfillment_decision_with_options(
+        self,
+        request: yunjian_20211217_models.AcceptFulfillmentDecisionRequest,
+        headers: Dict[str, str],
+        runtime: util_models.RuntimeOptions,
+    ) -> yunjian_20211217_models.AcceptFulfillmentDecisionResponse:
+        """
+        @summary 云产品交付决策反馈回调
+        
+        @param request: AcceptFulfillmentDecisionRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: AcceptFulfillmentDecisionResponse
+        """
+        UtilClient.validate_model(request)
+        body = {}
+        if not UtilClient.is_unset(request.decision_conclusion):
+            body['DecisionConclusion'] = request.decision_conclusion
+        if not UtilClient.is_unset(request.decision_type):
+            body['DecisionType'] = request.decision_type
+        if not UtilClient.is_unset(request.order_id):
+            body['OrderId'] = request.order_id
+        req = open_api_models.OpenApiRequest(
+            headers=headers,
+            body=OpenApiUtilClient.parse_to_map(body)
+        )
+        params = open_api_models.Params(
+            action='acceptFulfillmentDecision',
+            version='2021-12-17',
+            protocol='HTTPS',
+            pathname=f'/acceptFulfillmentDecision',
+            method='POST',
+            auth_type='AK',
+            style='ROA',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            yunjian_20211217_models.AcceptFulfillmentDecisionResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def accept_fulfillment_decision_with_options_async(
+        self,
+        request: yunjian_20211217_models.AcceptFulfillmentDecisionRequest,
+        headers: Dict[str, str],
+        runtime: util_models.RuntimeOptions,
+    ) -> yunjian_20211217_models.AcceptFulfillmentDecisionResponse:
+        """
+        @summary 云产品交付决策反馈回调
+        
+        @param request: AcceptFulfillmentDecisionRequest
+        @param headers: map
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: AcceptFulfillmentDecisionResponse
+        """
+        UtilClient.validate_model(request)
+        body = {}
+        if not UtilClient.is_unset(request.decision_conclusion):
+            body['DecisionConclusion'] = request.decision_conclusion
+        if not UtilClient.is_unset(request.decision_type):
+            body['DecisionType'] = request.decision_type
+        if not UtilClient.is_unset(request.order_id):
+            body['OrderId'] = request.order_id
+        req = open_api_models.OpenApiRequest(
+            headers=headers,
+            body=OpenApiUtilClient.parse_to_map(body)
+        )
+        params = open_api_models.Params(
+            action='acceptFulfillmentDecision',
+            version='2021-12-17',
+            protocol='HTTPS',
+            pathname=f'/acceptFulfillmentDecision',
+            method='POST',
+            auth_type='AK',
+            style='ROA',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            yunjian_20211217_models.AcceptFulfillmentDecisionResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def accept_fulfillment_decision(
+        self,
+        request: yunjian_20211217_models.AcceptFulfillmentDecisionRequest,
+    ) -> yunjian_20211217_models.AcceptFulfillmentDecisionResponse:
+        """
+        @summary 云产品交付决策反馈回调
+        
+        @param request: AcceptFulfillmentDecisionRequest
+        @return: AcceptFulfillmentDecisionResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        headers = {}
+        return self.accept_fulfillment_decision_with_options(request, headers, runtime)
+
+    async def accept_fulfillment_decision_async(
+        self,
+        request: yunjian_20211217_models.AcceptFulfillmentDecisionRequest,
+    ) -> yunjian_20211217_models.AcceptFulfillmentDecisionResponse:
+        """
+        @summary 云产品交付决策反馈回调
+        
+        @param request: AcceptFulfillmentDecisionRequest
+        @return: AcceptFulfillmentDecisionResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        headers = {}
+        return await self.accept_fulfillment_decision_with_options_async(request, headers, runtime)